refactor and decouple transpiler related codes
